+++ test/offline/apps/fpm_demo_base/fpm_demo_base/public/js/fpm_demo_base_lib.js
@@ -1,0 +1,2 @@
+export const marker = "FPM_DEMO_BASE_MARKER_LIB";
+window.fpm_demo_base_lib = marker;